+++ src/Components/WebView/Platforms/WindowsForms/src/WindowsFormsWebView2Wrapper.cs
@@ -2,7 +2,6 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.IO;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Components.WebView.WebView2;
 using Microsoft.Web.WebView2.Core;
@@ -57,201 +56,6 @@ namespace Microsoft.AspNetCore.Components.WebView.WindowsForms
         public Task EnsureCoreWebView2Async()
         {
             return WebView2.EnsureCoreWebView2Async(Environment);
-        }
-    }
-
-    internal class WindowsFormsCoreWebView2AcceleratorKeyPressedEventArgsWrapper : ICoreWebView2AcceleratorKeyPressedEventArgsWrapper
-    {
-        private readonly CoreWebView2AcceleratorKeyPressedEventArgs _eventArgs;
-
-        public WindowsFormsCoreWebView2AcceleratorKeyPressedEventArgsWrapper(CoreWebView2AcceleratorKeyPressedEventArgs eventArgs)
-        {
-            _eventArgs = eventArgs;
-        }
-        public uint VirtualKey => _eventArgs.VirtualKey;
-
-        public int KeyEventLParam => _eventArgs.KeyEventLParam;
-
-        public bool Handled
-        {
-            get => _eventArgs.Handled;
-            set => _eventArgs.Handled = value;
-        }
-    }
-
-    internal class WindowsFormsCoreWebView2Wrapper : ICoreWebView2Wrapper
-    {
-        private readonly WindowsFormsWebView2Wrapper _webView2;
-        private WindowsFormsCoreWebView2SettingsWrapper _settings;
-
-        public WindowsFormsCoreWebView2Wrapper(WindowsFormsWebView2Wrapper webView2)
-        {
-            if (webView2 is null)
-            {
-                throw new ArgumentNullException(nameof(webView2));
-            }
-
-            _webView2 = webView2;
-        }
-
-        public ICoreWebView2SettingsWrapper Settings
-        {
-            get
-            {
-                if (_settings == null)
-                {
-                    _settings = new WindowsFormsCoreWebView2SettingsWrapper(_webView2.WebView2.CoreWebView2.Settings);
-                }
-                return _settings;
-            }
-        }
-
-        public Task AddScriptToExecuteOnDocumentCreatedAsync(string javaScript)
-        {
-            return _webView2.WebView2.CoreWebView2.AddScriptToExecuteOnDocumentCreatedAsync(javaScript);
-        }
-
-        public Action AddWebMessageReceivedHandler(Action<WebMessageReceivedEventArgs> messageReceivedHandler)
-        {
-            void nativeEventHandler(object sender, CoreWebView2WebMessageReceivedEventArgs e)
-            {
-                messageReceivedHandler(new WebMessageReceivedEventArgs(e.Source, e.TryGetWebMessageAsString()));
-            }
-
-            _webView2.WebView2.CoreWebView2.WebMessageReceived += nativeEventHandler;
-
-            // Return removal callback
-            return () =>
-            {
-                _webView2.WebView2.CoreWebView2.WebMessageReceived -= nativeEventHandler;
-            };
-        }
-
-        public void AddWebResourceRequestedFilter(string uri, CoreWebView2WebResourceContextWrapper resourceContext)
-        {
-            _webView2.WebView2.CoreWebView2.AddWebResourceRequestedFilter(uri, (CoreWebView2WebResourceContext)resourceContext);
-        }
-
-        public Action AddWebResourceRequestedHandler(EventHandler<ICoreWebView2WebResourceRequestedEventArgsWrapper> webResourceRequestedHandler)
-        {
-            void nativeEventHandler(object sender, CoreWebView2WebResourceRequestedEventArgs e)
-            {
-                webResourceRequestedHandler(_webView2.WebView2, new WindowsFormsCoreWebView2WebResourceRequestedEventArgsWrapper(_webView2.Environment, e));
-            }
-
-            _webView2.WebView2.CoreWebView2.WebResourceRequested += nativeEventHandler;
-
-            // Return removal callback
-            return () =>
-            {
-                _webView2.WebView2.CoreWebView2.WebResourceRequested -= nativeEventHandler;
-            };
-        }
-
-        public void PostWebMessageAsString(string webMessageAsString)
-        {
-            _webView2.WebView2.CoreWebView2.PostWebMessageAsString(webMessageAsString);
-        }
-    }
-
-    internal class WindowsFormsCoreWebView2SettingsWrapper : ICoreWebView2SettingsWrapper
-    {
-        private readonly CoreWebView2Settings _settings;
-
-        public WindowsFormsCoreWebView2SettingsWrapper(CoreWebView2Settings settings)
-        {
-            _settings = settings;
-        }
-
-        public bool IsScriptEnabled
-        {
-            get => _settings.IsScriptEnabled;
-            set => _settings.IsScriptEnabled = value;
-        }
-        public bool IsWebMessageEnabled
-        {
-            get => _settings.IsWebMessageEnabled;
-            set => _settings.IsWebMessageEnabled = value;
-        }
-        public bool AreDefaultScriptDialogsEnabled
-        {
-            get => _settings.AreDefaultScriptDialogsEnabled;
-            set => _settings.AreDefaultScriptDialogsEnabled = value;
-        }
-        public bool IsStatusBarEnabled
-        {
-            get => _settings.IsStatusBarEnabled;
-            set => _settings.IsStatusBarEnabled = value;
-        }
-        public bool AreDevToolsEnabled
-        {
-            get => _settings.AreDevToolsEnabled;
-            set => _settings.AreDevToolsEnabled = value;
-        }
-        public bool AreDefaultContextMenusEnabled
-        {
-            get => _settings.AreDefaultContextMenusEnabled;
-            set => _settings.AreDefaultContextMenusEnabled = value;
-        }
-        public bool AreHostObjectsAllowed
-        {
-            get => _settings.AreHostObjectsAllowed;
-            set => _settings.AreHostObjectsAllowed = value;
-        }
-        public bool IsZoomControlEnabled
-        {
-            get => _settings.IsZoomControlEnabled;
-            set => _settings.IsZoomControlEnabled = value;
-        }
-        public bool IsBuiltInErrorPageEnabled
-        {
-            get => _settings.IsBuiltInErrorPageEnabled;
-            set => _settings.IsBuiltInErrorPageEnabled = value;
-        }
-    }
-
-    internal class WindowsFormsCoreWebView2WebResourceRequestedEventArgsWrapper : ICoreWebView2WebResourceRequestedEventArgsWrapper
-    {
-        private readonly CoreWebView2Environment _environment;
-        private readonly CoreWebView2WebResourceRequestedEventArgs _webResourceRequestedEventArgs;
-
-        public WindowsFormsCoreWebView2WebResourceRequestedEventArgsWrapper(CoreWebView2Environment environment, CoreWebView2WebResourceRequestedEventArgs webResourceRequestedEventArgs)
-        {
-            _environment = environment;
-            _webResourceRequestedEventArgs = webResourceRequestedEventArgs;
-
-            Request = new WindowsFormsCoreWebView2WebResourceRequestWrapper(webResourceRequestedEventArgs);
-            ResourceContext = (CoreWebView2WebResourceContextWrapper)webResourceRequestedEventArgs.ResourceContext;
-        }
-
-        public ICoreWebView2WebResourceRequestWrapper Request { get; }
-
-        public CoreWebView2WebResourceContextWrapper ResourceContext { get; }
-
-        public void SetResponse(Stream content, int statusCode, string statusMessage, string headerString)
-        {
-            _webResourceRequestedEventArgs.Response = _environment.CreateWebResourceResponse(content, statusCode, statusMessage, headerString);
-        }
-    }
-
-    internal class WindowsFormsCoreWebView2WebResourceRequestWrapper : ICoreWebView2WebResourceRequestWrapper
-    {
-        private readonly CoreWebView2WebResourceRequestedEventArgs _webResourceRequestedEventArgs;
-
-        public WindowsFormsCoreWebView2WebResourceRequestWrapper(CoreWebView2WebResourceRequestedEventArgs webResourceRequestedEventArgs)
-        {
-            _webResourceRequestedEventArgs = webResourceRequestedEventArgs;
-        }
-
-        public string Uri
-        {
-            get => _webResourceRequestedEventArgs.Request.Uri;
-            set => _webResourceRequestedEventArgs.Request.Uri = value;
-        }
-        public string Method
-        {
-            get => _webResourceRequestedEventArgs.Request.Method;
-            set => _webResourceRequestedEventArgs.Request.Method = value;
         }
     }
 }
